Split the codebase to make it more modular.
- added spif_util.h to hold common structs and enums.
- ported the register defintions to source files to prevent conflicts.
- changed the functions in spif_compat_is25xx and spif_compat_w25xx to use correct registers.
