+++ spif.c
@@ -5,7 +5,6 @@
 
 #include "spif.h"
 #include <stdlib.h>
-#include "NimaLTD.I-CUBE-SPIF_conf.h"
 
 
 #if SPIF_RTOS == SPIF_RTOS_DISABLE
@@ -18,21 +17,6 @@
 #elif SPIF_RTOS == SPIF_RTOS_THREADX
 #include "app_threadx.h"
 #endif
-
-
-#if (SPIF_COMPAT == SPIF_COMPAT_W25XX)
-  #include "spif_compat_w25xx.h"
-#elif (SPIF_COMPAT == SPIF_COMPAT_IS25XX)
-  #include "spif_compat_is25xx.h"
-#endif
-
-#if (SPIF_PLATFORM == SPIF_PLATFORM_OCTOSPI)
-  #include "spif_interface_octospi.h"
-#else
-  #include "spif_interface_spi.h"
-#endif
-
-
 
 /***********************************************************************************************************/
 
@@ -79,540 +63,20 @@ void SPIF_CsPin(SPIF_HandleTypeDef *Handle, bool Select)
   for (int i = 0; i < 10; i++);
 }
 
-/***********************************************************************************************************/
-
-bool SPIF_WriteEnable(SPIF_HandleTypeDef *Handle)
-{
-  bool retVal = true;
-  uint8_t tx[1] = {SPIF_CMD_WRITEENABLE};
-  SPIF_CsPin(Handle, 0);
-  if (SPIF_Transmit(Handle, tx, 1, 100) == false)
-  {
-    retVal = false;
-    dprintf("SPIF_WriteEnable() Error\r\n");
-  }
-  SPIF_CsPin(Handle, 1);
-  return retVal;
-}
-
-/***********************************************************************************************************/
-
-bool SPIF_WriteDisable(SPIF_HandleTypeDef *Handle)
-{
-  bool retVal = true;
-  uint8_t tx[1] = {SPIF_CMD_WRITEDISABLE};
-  SPIF_CsPin(Handle, 0);
-  if (SPIF_Transmit(Handle, tx, 1, 100) == false)
-  {
-    retVal = false;
-    dprintf("SPIF_WriteDisable() Error\r\n");
-  }
-  SPIF_CsPin(Handle, 1);
-  return retVal;
-}
-
-/***********************************************************************************************************/
-
-bool SPIF_WaitForWriting(SPIF_HandleTypeDef *Handle, uint32_t Timeout)
-{
-  bool retVal = false;
-  uint32_t startTime = HAL_GetTick();
-  while (1)
-  {
-    SPIF_Delay(1);
-    if (HAL_GetTick() - startTime >= Timeout)
-    {
-      dprintf("SPIF_WaitForWriting() TIMEOUT\r\n");
-      break;
-    }
-    #if (SPIF_COMPAT == SPIF_COMPAT_W25XX)
-    if ((SPIF_W25XX_ReadReg1(Handle) & SPIF_STATUS1_BUSY) == 0)
-    #elif (SPIF_COMPAT == SPIF_COMPAT_IS25XX)
-    if ((SPIF_IS25_ReadReg1(Handle) & SPIF_STATUS_WIP) == 0)
-    #endif
-    {
-      retVal = true;
-      break;
-    }
-  }
-  return retVal;
-}
-
-/***********************************************************************************************************/
-
-bool SPIF_FindChip(SPIF_HandleTypeDef *Handle)
-{
-  uint8_t tx[4] = {SPIF_CMD_JEDECID, 0xFF, 0xFF, 0xFF};
-  uint8_t rx[4];
-  bool retVal = false;
-  do
-  {
-    dprintf("SPIF_FindChip()\r\n");
-    SPIF_CsPin(Handle, 0);
-    if (SPIF_TransmitReceive(Handle, tx, rx, 4, 100) == false)
-    {
-      SPIF_CsPin(Handle, 1);
-      break;
-    }
-    SPIF_CsPin(Handle, 1);
-    dprintf("CHIP ID: 0x%02X%02X%02X\r\n", rx[1], rx[2], rx[3]);
-    Handle->Manufactor = rx[1];
-    Handle->MemType = rx[2];
-    Handle->Size = rx[3];
-
-    dprintf("SPIF MANUFACTURE: ");
-    switch (Handle->Manufactor)
-    {
-    case SPIF_MANUFACTOR_WINBOND:
-      dprintf("WINBOND");
-      break;
-    case SPIF_MANUFACTOR_SPANSION:
-      dprintf("SPANSION");
-      break;
-    case SPIF_MANUFACTOR_MICRON:
-      dprintf("MICRON");
-      break;
-    case SPIF_MANUFACTOR_MACRONIX:
-      dprintf("MACRONIX");
-      break;
-    case SPIF_MANUFACTOR_ISSI:
-      dprintf("ISSI");
-      break;
-    case SPIF_MANUFACTOR_GIGADEVICE:
-      dprintf("GIGADEVICE");
-      break;
-    case SPIF_MANUFACTOR_AMIC:
-      dprintf("AMIC");
-      break;
-    case SPIF_MANUFACTOR_SST:
-      dprintf("SST");
-      break;
-    case SPIF_MANUFACTOR_HYUNDAI:
-      dprintf("HYUNDAI");
-      break;
-    case SPIF_MANUFACTOR_FUDAN:
-      dprintf("FUDAN");
-      break;
-    case SPIF_MANUFACTOR_ESMT:
-      dprintf("ESMT");
-      break;
-    case SPIF_MANUFACTOR_INTEL:
-      dprintf("INTEL");
-      break;
-    case SPIF_MANUFACTOR_SANYO:
-      dprintf("SANYO");
-      break;
-    case SPIF_MANUFACTOR_FUJITSU:
-      dprintf("FUJITSU");
-      break;
-    case SPIF_MANUFACTOR_EON:
-      dprintf("EON");
-      break;
-    case SPIF_MANUFACTOR_PUYA:
-      dprintf("PUYA");
-      break;
-    default:
-      Handle->Manufactor = SPIF_MANUFACTOR_ERROR;
-      dprintf("ERROR");
-      break;
-    }
-    dprintf(" - MEMTYPE: 0x%02X", Handle->MemType);
-    dprintf(" - SIZE: ");
-    switch (Handle->Size)
-    {
-    case SPIF_SIZE_1MBIT:
-      Handle->BlockCnt = 2;
-      dprintf("1 MBIT\r\n");
-      break;
-    case SPIF_SIZE_2MBIT:
-      Handle->BlockCnt = 4;
-      dprintf("2 MBIT\r\n");
-      break;
-    case SPIF_SIZE_4MBIT:
-      Handle->BlockCnt = 8;
-      dprintf("4 MBIT\r\n");
-      break;
-    case SPIF_SIZE_8MBIT:
-      Handle->BlockCnt = 16;
-      dprintf("8 MBIT\r\n");
-      break;
-    case SPIF_SIZE_16MBIT:
-      Handle->BlockCnt = 32;
-      dprintf("16 MBIT\r\n");
-      break;
-    case SPIF_SIZE_32MBIT:
-      Handle->BlockCnt = 64;
-      dprintf("32 MBIT\r\n");
-      break;
-    case SPIF_SIZE_64MBIT:
-      Handle->BlockCnt = 128;
-      dprintf("64 MBIT\r\n");
-      break;
-    case SPIF_SIZE_128MBIT:
-      Handle->BlockCnt = 256;
-      dprintf("128 MBIT\r\n");
-      break;
-    case SPIF_SIZE_256MBIT:
-      Handle->BlockCnt = 512;
-      dprintf("256 MBIT\r\n");
-      break;
-    case SPIF_SIZE_512MBIT:
-      Handle->BlockCnt = 1024;
-      dprintf("512 MBIT\r\n");
-      break;
-    default:
-      Handle->Size = SPIF_SIZE_ERROR;
-      dprintf("ERROR\r\n");
-      break;
-    }
-
-    Handle->SectorCnt = Handle->BlockCnt * 16;
-    Handle->PageCnt = (Handle->SectorCnt * SPIF_SECTOR_SIZE) / SPIF_PAGE_SIZE;
-    dprintf("SPIF BLOCK CNT: %ld\r\n", Handle->BlockCnt);
-    dprintf("SPIF SECTOR CNT: %ld\r\n", Handle->SectorCnt);
-    dprintf("SPIF PAGE CNT: %ld\r\n", Handle->PageCnt);
-    dprintf("SPIF STATUS1: 0x%02X\r\n", SPIF_ReadReg1(Handle));
-    dprintf("SPIF STATUS2: 0x%02X\r\n", SPIF_ReadReg2(Handle));
-    dprintf("SPIF STATUS3: 0x%02X\r\n", SPIF_ReadReg3(Handle));
-    retVal = true;
-
-  } while (0);
-
-  return retVal;
-}
-
-
-bool SPIF_TransmitReceive(SPIF_HandleTypeDef *Handle, uint8_t *Tx, uint8_t *Rx, size_t Size, uint32_t Timeout){
-#if (SPIF_PLATFORM == SPIF_PLATFORM_OCTOSPI)
-	return SPIF_OCTOSPI_TransmitReceive(Handle, Tx, Rx, Size, Timeout);
-#else
-	return SPIF_SPI_TransmitReceive(Handle, Tx, Rx, Size, Timeout);
-#endif
-}
-
-bool SPIF_Transmit(SPIF_HandleTypeDef *Handle, uint8_t *Tx, size_t Size, uint32_t Timeout){
-#if (SPIF_PLATFORM == SPIF_PLATFORM_OCTOSPI)
-	return SPIF_OCTOSPI_Transmit(Handle, Tx, Size, Timeout);
-#else
-	return SPIF_SPI_Transmit(Handle, Tx, Size, Timeout);
-#endif
-}
-
-bool SPIF_Receive(SPIF_HandleTypeDef *Handle, uint8_t *Rx, size_t Size, uint32_t Timeout){
-#if (SPIF_PLATFORM == SPIF_PLATFORM_OCTOSPI)
-	return SPIF_OCTOSPI_Receive(Handle, Rx, Size, Timeout);
-#else
-	return SPIF_SPI_Receive(Handle, Rx, Size, Timeout);
-#endif
-}
 
 
 /***********************************************************************************************************/
 
-bool SPIF_WriteFn(SPIF_HandleTypeDef *Handle, uint32_t PageNumber, uint8_t *Data, uint32_t Size, uint32_t Offset)
-{
-  bool retVal = false;
-  uint32_t address = 0, maximum = SPIF_PAGE_SIZE - Offset;
-  uint8_t tx[5];
-  do
-  {
-#if SPIF_DEBUG != SPIF_DEBUG_DISABLE
-    uint32_t dbgTime = HAL_GetTick();
-#endif
-    dprintf("SPIF_WritePage() START PAGE %ld\r\n", PageNumber);
-    if (PageNumber >= Handle->PageCnt)
-    {
-      dprintf("SPIF_WritePage() ERROR PageNumber\r\n");
-      break;
-    }
-    if (Offset >= SPIF_PAGE_SIZE)
-    {
-      dprintf("SPIF_WritePage() ERROR Offset\r\n");
-      break;
-    }
-    if (Size > maximum)
-    {
-      Size = maximum;
-    }
-    address = SPIF_PageToAddress(PageNumber) + Offset;
-#if SPIF_DEBUG == SPIF_DEBUG_FULL
-      dprintf("SPIF WRITING {\r\n0x%02X", Data[0]);
-      for (int i = 1; i < Size; i++)
-      {
-        if (i % 8 == 0)
-        {
-          dprintf("\r\n");
-        }
-        dprintf(", 0x%02X", Data[i]);
-      }
-      dprintf("\r\n}\r\n");
-#endif
-    if (SPIF_WriteEnable(Handle) == false)
-    {
-      break;
-    }
-    SPIF_CsPin(Handle, 0);
-    if (Handle->BlockCnt >= 512)
-    {
-      tx[0] = SPIF_CMD_PAGEPROG4ADD;
-      tx[1] = (address & 0xFF000000) >> 24;
-      tx[2] = (address & 0x00FF0000) >> 16;
-      tx[3] = (address & 0x0000FF00) >> 8;
-      tx[4] = (address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 5, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    else
-    {
-      tx[0] = SPIF_CMD_PAGEPROG3ADD;
-      tx[1] = (address & 0x00FF0000) >> 16;
-      tx[2] = (address & 0x0000FF00) >> 8;
-      tx[3] = (address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 4, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    if (SPIF_Transmit(Handle, Data, Size, 1000) == false)
-    {
-      SPIF_CsPin(Handle, 1);
-      break;
-    }
-    SPIF_CsPin(Handle, 1);
-    if (SPIF_WaitForWriting(Handle, 100))
-    {
-      dprintf("SPIF_WritePage() %d BYTES WITERN DONE AFTER %ld ms\r\n", (uint16_t)Size, HAL_GetTick() - dbgTime);
-      retVal = true;
-    }
 
-  } while (0);
 
-  SPIF_WriteDisable(Handle);
-  return retVal;
-}
 
 /***********************************************************************************************************/
-
-bool SPIF_ReadFn(SPIF_HandleTypeDef *Handle, uint32_t Address, uint8_t *Data, uint32_t Size)
-{
-  bool retVal = false;
-  uint8_t tx[5];
-  do
-  {
-#if SPIF_DEBUG != SPIF_DEBUG_DISABLE
-    uint32_t dbgTime = HAL_GetTick();
-#endif
-    dprintf("SPIF_ReadAddress() START ADDRESS %ld\r\n", Address);
-    SPIF_CsPin(Handle, 0);
-    if (Handle->BlockCnt >= 512)
-    {
-      tx[0] = SPIF_CMD_READDATA4ADD;
-      tx[1] = (Address & 0xFF000000) >> 24;
-      tx[2] = (Address & 0x00FF0000) >> 16;
-      tx[3] = (Address & 0x0000FF00) >> 8;
-      tx[4] = (Address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 5, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    else
-    {
-      tx[0] = SPIF_CMD_READDATA3ADD;
-      tx[1] = (Address & 0x00FF0000) >> 16;
-      tx[2] = (Address & 0x0000FF00) >> 8;
-      tx[3] = (Address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 4, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    uint16_t max_uint16_size = 0xFFFF - 2;
-    if(Size > max_uint16_size)
-    {
-    	uint32_t delta = Size;
-    	uint32_t Data_offset = 0;
-    	while(delta > 0)
-    	{
-    		if(delta <= max_uint16_size)
-    		{
-    			if (SPIF_Receive(Handle, (Data+Data_offset), delta, 2000) == false)
-    			{
-    				SPIF_CsPin(Handle, 1);
-    				break;
-    			}
-    			delta = 0;
-    		}
-    		else
-    		{
-    			if (SPIF_Receive(Handle, (Data+Data_offset), max_uint16_size, 2000) == false)
-    			{
-    				SPIF_CsPin(Handle, 1);
-    				break;
-    			}
-    			delta = abs(delta - max_uint16_size);
-    			Data_offset += max_uint16_size;
-    		}
-    	}
-    	if(delta > 0)
-    	{
-    		//break the main do-while loop, so as to return false
-    		break;
-    	}
-    }
-    else
-    {
-    	if(SPIF_Receive(Handle, Data, Size, 2000) == false)
-    	{
-    		SPIF_CsPin(Handle, 1);
-    		break;
-    	}
-    }
-    SPIF_CsPin(Handle, 1);
-    dprintf("SPIF_ReadAddress() %d BYTES READ DONE AFTER %ld ms\r\n", (uint16_t)Size, HAL_GetTick() - dbgTime);
-#if SPIF_DEBUG == SPIF_DEBUG_FULL
-    dprintf("{\r\n0x%02X", Data[0]);
-    for (int i = 1; i < Size; i++)
-    {
-      if (i % 8 == 0)
-      {
-        dprintf("\r\n");
-      }
-      dprintf(", 0x%02X", Data[i]);
-    }
-    dprintf("\r\n}\r\n");
-#endif
-    retVal = true;
-
-  } while (0);
-
-  return retVal;
-}
 
 /************************************************************************************************************
 **************    Public Functions
 ************************************************************************************************************/
 
-/**
-  * @brief  Full Erase chip.
-  * @note   Send the Full-Erase-chip command and wait for completion
-  *
-  * @param  *Handle: Pointer to SPIF_HandleTypeDef structure
-  *
-  * @retval bool: true or false
-  */
-bool SPIF_EraseChip(SPIF_HandleTypeDef *Handle)
-{
-  SPIF_Lock(Handle);
-  bool retVal = false;
-  uint8_t tx[1] = {SPIF_CMD_CHIPERASE1};
-  do
-  {
-#if SPIF_DEBUG != SPIF_DEBUG_DISABLE
-    uint32_t dbgTime = HAL_GetTick();
-#endif
-    dprintf("SPIF_EraseChip() START\r\n");
-    if (SPIF_WriteEnable(Handle) == false)
-    {
-      break;
-    }
-    SPIF_CsPin(Handle, 0);
-    if (SPIF_Transmit(Handle, tx, 1, 100) == false)
-    {
-      SPIF_CsPin(Handle, 1);
-      break;
-    }
-    SPIF_CsPin(Handle, 1);
-    if (SPIF_WaitForWriting(Handle, Handle->BlockCnt * 1000))
-    {
-      dprintf("SPIF_EraseChip() DONE AFTER %ld ms\r\n", HAL_GetTick() - dbgTime);
-      retVal = true;
-    }
-
-  } while (0);
-
-  SPIF_WriteDisable(Handle);
-  SPIF_UnLock(Handle);
-  return retVal;
-}
-
 /***********************************************************************************************************/
-
-/**
-  * @brief  Erase Sector.
-  * @note   Send the Erase-Sector command and wait for completion
-  *
-  * @param  *Handle: Pointer to SPIF_HandleTypeDef structure
-  * @param  Sector: Selected Sector
-  *
-  * @retval bool: true or false
-  */
-bool SPIF_EraseSector(SPIF_HandleTypeDef *Handle, uint32_t Sector)
-{
-  SPIF_Lock(Handle);
-  bool retVal = false;
-  uint32_t address = Sector * SPIF_SECTOR_SIZE;
-  uint8_t tx[5];
-  do
-  {
-#if SPIF_DEBUG != SPIF_DEBUG_DISABLE
-    uint32_t dbgTime = HAL_GetTick();
-#endif
-    dprintf("SPIF_EraseSector() START SECTOR %ld\r\n", Sector);
-    if (Sector >= Handle->SectorCnt)
-    {
-      dprintf("SPIF_EraseSector() ERROR Sector NUMBER\r\n");
-      break;
-    }
-    if (SPIF_WriteEnable(Handle) == false)
-    {
-      break;
-    }
-    SPIF_CsPin(Handle, 0);
-    if (Handle->BlockCnt >= 512)
-    {
-      tx[0] = SPIF_CMD_SECTORERASE4ADD;
-      tx[1] = (address & 0xFF000000) >> 24;
-      tx[2] = (address & 0x00FF0000) >> 16;
-      tx[3] = (address & 0x0000FF00) >> 8;
-      tx[4] = (address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 5, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    else
-    {
-      tx[0] = SPIF_CMD_SECTORERASE3ADD;
-      tx[1] = (address & 0x00FF0000) >> 16;
-      tx[2] = (address & 0x0000FF00) >> 8;
-      tx[3] = (address & 0x000000FF);
-      if (SPIF_Transmit(Handle, tx, 4, 100) == false)
-      {
-        SPIF_CsPin(Handle, 1);
-        break;
-      }
-    }
-    SPIF_CsPin(Handle, 1);
-    if (SPIF_WaitForWriting(Handle, 1000))
-    {
-      dprintf("SPIF_EraseSector() DONE AFTER %ld ms\r\n", HAL_GetTick() - dbgTime);
-      retVal = true;
-    }
-
-  } while (0);
-
-  SPIF_WriteDisable(Handle);
-  SPIF_UnLock(Handle);
-  return retVal;
-}
 
 /***********************************************************************************************************/
 
@@ -907,32 +371,6 @@ bool SPIF_ReadBlock(SPIF_HandleTypeDef *Handle, uint32_t BlockNumber, uint8_t *D
   SPIF_UnLock(Handle);
   return retVal;
 }
-
-
-#if (SPIF_COMPAT == SPIF_COMPAT_IS25XX)
-
-/*
-  * @brief  Write QPI Enable command
-  * @note   Send the QPI-Enable command
-  * 
-  * @param  *Handle: Pointer to SPIF_HandleTypeDef structure
-  * 
-  * @retval bool: true or false
-  */
- bool SPIF_QPI_Enable(SPIF_HandleTypeDef *Handle)
- {
-   bool retVal = true;
-   uint8_t tx[1] = {SPIF_CMD_ENTERQPI};
-   SPIF_CsPin(Handle, 0);
-   if (SPIF_Transmit(Handle, tx, 1, 100) == false)
-   {
-     retVal = false;
-     dprintf("SPIF_QPI_Enable() Error\r\n");
-   }
-   SPIF_CsPin(Handle, 1);
-   return retVal;
- }
-#endif 
 
 /*
   * @brief  Send Command to SPIF
